pyproject.toml
[tool.poetry]
name = "mega-pints"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
  {include = "mega-pints.py"}
]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.11.1"
ImageHash = "^4.3.1"
Pillow = "^9.3.0"
selenium = "^4.7.2"
requests = "^2.28.1"
tdqm = "^0.0.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
pre-commit = "^2.17.0"
bandit = "^1.7.0"
isort = "^5.9.1"
pylint = "^2.12.2"
mypy = "^0.961"
twine = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

